+++ propagator/opentelemetry-propagator-aws-xray/tests/trace/propagation/test_aws_xray_format.py
@@ -19,7 +19,10 @@ from requests.structures import CaseInsensitiveDict
 
 import opentelemetry.trace as trace_api
 from opentelemetry.context import Context
-from opentelemetry.propagators.aws import TRACE_HEADER_KEY, AwsXRayFormat
+from opentelemetry.propagators.aws.aws_xray_propagator import (
+    TRACE_HEADER_KEY,
+    AwsXRayPropagator,
+)
 from opentelemetry.trace import (
     DEFAULT_TRACE_OPTIONS,
     DEFAULT_TRACE_STATE,
@@ -82,7 +85,7 @@ def build_test_span_context(
 
 
 class AwsXRayPropagatorTest(unittest.TestCase):
-    XRAY_PROPAGATOR = AwsXRayFormat()
+    XRAY_PROPAGATOR = AwsXRayPropagator()
 
     # Inject Tests
 
@@ -296,7 +299,7 @@ class AwsXRayPropagatorTest(unittest.TestCase):
 
                 self.assertDictEqual(Context(), ctx)
 
-    @patch("opentelemetry.propagators.aws.trace")
+    @patch("opentelemetry.propagators.aws.aws_xray_propagator.trace")
     def test_fields(self, mock_trace):
         """Make sure the fields attribute returns the fields used in inject"""